Lua API: wesnoth.map.parse|make_bitmap are now defined in the base kernel
This is mainly so that they can be used in the mapgen kernel; exposing them to the plugin kernel is just an incidental side-effect.
